changes to fix issue 55 as well as restructuring
