--- client/src/pages/GeneralInfo/GeneralInfo.module.scss
@@ -1,22 +1,0 @@
-.lato-regular {
-    font-family: "Lato", sans-serif;
-    font-weight: 400;
-    font-style: normal;
-  }
-  
-  .lato-bold {
-    font-family: "Lato", sans-serif;
-    font-weight: 600;
-    font-style: normal;
-  }  
-
-.GeneralInfo {
-    font-family: 'Lato', sans-serif;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    height: 100%;
-    width: 100%;
-    padding: 0 20px;
-}
