+++ client/src/pages/SignUp/GeneralInfo/GeneralInfo.module.scss
@@ -1,0 +1,48 @@
+.lato-regular {
+  font-family: "Lato", sans-serif;
+  font-weight: 400;
+  font-style: normal;
+}
+
+.lato-bold {
+  font-family: "Lato", sans-serif;
+  font-weight: 600;
+  font-style: normal;
+}
+
+.GeneralInfo {
+  &__Container {
+    display: flex;
+    flex-direction: column;
+    align-items: center; /* Center horizontally */
+    justify-content: center; /* Center vertically */
+    min-height: 100vh; /* Adjust height as needed */
+    background-color: #f0f0f0;
+  }
+  &__whiteContainer {
+    background-color: white; /* Set the background color of the container */
+    padding: 20px; /* Add padding to create space between the content and the container */
+    border-radius: 10px; /* Add rounded corners to the container */
+    width: 80%; /* Adjust the width of the container */
+    max-width: 280px; /* Set maximum width for responsiveness */
+    min-height: 600px;
+  }
+  &__textCentered {
+    text-align: center;
+  }
+  &__continueHereContainer {
+    display: flex;
+    margin-top: 40px;
+    justify-content: center; /* Center vertically */
+  }
+  &__continueHereBtn {
+    color: green;
+  }
+
+  @media (max-height: 667px) {
+    &__whiteContainer {
+      min-height: 80vh;
+      padding: 10px;
+    }
+  }
+}
