+++ client/src/pages/SignUp/GeneralInfo/GeneralInfo.tsx
@@ -1,9 +1,10 @@
 import styles from './GeneralInfo.module.scss';
 import { useState } from 'react';
 import { Button, Box, Stack, Typography } from '@mui/material';
-import { TextField } from '@/components';
+import { TextField, CustomButton } from '@/components';
 
 function GeneralInfo(): JSX.Element {
+  // State variables to store form input values
   const [firstName, setFirstName] = useState<string>('');
   const [lastName, setLastName] = useState<string>('');
   const [userName, setUserName] = useState<string>('');
@@ -11,16 +12,20 @@ function GeneralInfo(): JSX.Element {
   const [phone, setPhone] = useState<string>('');
   const [DoB, setDoB] = useState<string>('');
 
+  // Function to handle button click
+  const handleClick = () => {
+    console.log('Button clicked!');
+  };
+
   return (
-    <div className={styles['GeneralInfo']}>
-      <Box sx={{ width: '100%', maxWidth: 500 }}>
-        <Typography variant="h4" gutterBottom align="center">
+    <div className={styles['GeneralInfo__Container']}>
+      <Box className={styles['GeneralInfo__whiteContainer']} sx={{ width: '75%', maxWidth: 500}}>
+        <Typography variant="h4" gutterBottom align="center" sx={{ paddingBottom: '20px' }}>
           General Info
         </Typography>
-      </Box>
 
-      <Box sx={{ width: '100%', maxWidth: 500 }}>
         <Stack spacing={2}>
+          {/* Text fields for user input */}
           <TextField header="First Name" setInputValue={setFirstName} type="text" />
           <TextField header="Last Name" setInputValue={setLastName} type="text" />
           <TextField header="Username" setInputValue={setUserName} type="text" />
@@ -28,10 +33,11 @@ function GeneralInfo(): JSX.Element {
           <TextField header="Phone" setInputValue={setPhone} type="tel" />
           <TextField header="Date of Birth" setInputValue={setDoB} type="date" />
 
-          <Box sx={{ mt: 2, display: 'flex', justifyContent: 'center' }}>
-            <Button variant="contained" size="large" color="success">
+          <Box className={styles['GeneralInfo__continueHereContainer']} sx={{ mt: 2, display: 'flex', justifyContent: 'center' }}>
+            {/* Custom button for form submission */}
+            <CustomButton className={styles['GeneralInfo__continueHereBtn']} onClick={handleClick}>
               Continue
-            </Button>
+            </CustomButton>
           </Box>
         </Stack>
       </Box>
