+++ client/src/pages/index.ts
@@ -16,5 +16,5 @@ export * from './Profile/Profile';
 export { default as SignUp } from './SignUp/SignUp';
 export * from './SignUp/SignUp';
 
-export { default as GeneralInfo } from './GeneralInfo/GeneralInfo';
-export * from './GeneralInfo/GeneralInfo';
+export { default as GeneralInfo } from './SignUp/GeneralInfo/GeneralInfo';
+export * from './SignUp/GeneralInfo/GeneralInfo';
